Feature/profile/hj

See merge request s08-bigdata-recom-sub2/S08P22E202!47

Former-commit-id: 44982eb2cd8110d78b4444a22751cc0917316da8 [formerly d55f64e06ad8d89f77dc8b3f7efdef5e5fa12312]
Former-commit-id: 1302108c5a4fb996ed5516f0c8d3be55d529f181

+++ back-server/src/main/java/com/ssafy/ilbok/Repository/CitiesRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CitiesRepository extends JpaRepository<Cities, Long> {
+public interface CitiesRepository extends JpaRepository<Cities, Integer> {
     List<Cities> findByRegionCode(int regionCode);
 }
